+++ tests/test_log.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # FoFiX
-# Copyright (C) 2017 FoFiX team
+# Copyright (C) 2017-2018 FoFiX team
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -33,6 +33,7 @@ class ConfigureTest(unittest.TestCase):
 
     def test_logger(self):
         # configurer the logger
+        #log.configure(self.logfile.name, logging.INFO)
         log.configure(self.logfile.name)
         self.assertTrue(os.path.exists(self.logfile.name))
 
@@ -45,3 +46,26 @@ class ConfigureTest(unittest.TestCase):
         log_msg = self.logfile.readline().decode()
         self.assertIn('INFO', log_msg)
         self.assertIn(msg, log_msg)
+
+    def test_logger_warning_level(self):
+        # configurer the logger in Warning level
+        #log.configure(self.logfile.name, logging.WARNING)
+        log.configure(self.logfile.name, 'WARNING')
+
+        # log two messages
+        logger = logging.getLogger(__name__)
+        msg_info = "Test info log"
+        msg_warning = "Test warning log"
+        logger.info(msg_info)
+        logger.warning(msg_warning)
+
+        # read the log file
+        log_msgs = self.logfile.readlines()
+        self.assertEqual(len(log_msgs), 1)
+        log_msg = log_msgs[0].decode()
+
+        # check the log file content
+        self.assertNotIn('INFO', log_msg)
+        self.assertNotIn(msg_info, log_msg)
+        self.assertIn('WARNING', log_msg)
+        self.assertIn(msg_warning, log_msg)