+++ doc/source/releasenotes.rst
@@ -5,6 +5,7 @@ Release notes
 ------------------
 
 - Pin versions
+- Make the log level customizable
 
 
 0.3.0 (2017-11-09)